chore: fix folder name from Capitalize, rename twice, first one
