+++ src/components/_navList/NavList.scss
